Move prefix into const
